Tournament Manager fully working
Tournament manager can now play 100 matches between two executables very quickly. It gives the expected results: the TT boosts performance, and so does a longer search.

I still need to add in a standard database of FEN strings that can be used for opening, middle-game & endgame positions. As well as this, consider some variables like match length.

I still need to experiment with the TT and consider what nodes we store. I've created the object SearchParameters which sets all the constants for the search, such as TT size. This has allowed the perft state to be created with an empty TT (as it's not needed).

Also the program can now successfully create TTs of 100mb. So I should probably be much more aggressive in storing nodes. Are my Zobrist keys optimal enough?