Still fucked
{Changed to Fail Soft alpha/beta framework. This means that we can return scores that can be outside of the (alpha, beta) window. Instead of returning alpha in the case of fail low, and beta in the case of fail high}

The TT is detrimental to performance in both the QSearch and the Regular search.

I've found stripped back the TT to only store exact evaluations, and only in the regular search. I also limited probing to an identical depth, but carried out the search without using the TT move. I then compared the results of the search, to the results of the TT. In every single case, they were as expected. A lower bound was lower, an upper bound was higher, an exact bound was the same. And exact bounds returned the same move.

The decrease in performance could be due to: (i think it's a mix of both).
a. Time savings from considering exact entries don't cover TT overhead.
b. I am not handling upper/ lower bounds properly.