Okay I did something wrong
So now not so convinced by the TT, even when just used for move ordering it seems to have minor impact.

HOLY SHIT I FOUND A TINY LOGIC ERROR WITH CLASSIFYING node types using a combination of the orignial alpha beta and the current one. Now the TT is working fucking beautifully? Idk man I am just a genius

Commiting just to save it